refactor(common-data): системный каталог на общий стек полей, удалён systemFields (issue #73, шаг 1)
Устранён самый крупный единичный UI-дубль: SystemCommonDataPage держал приватную
копию стека рендера полей (common-data/systemFields.tsx, 527 стр.), дублируя
document-sets/fields/*. Переведён на общий стек (ComplexFieldGroup/ArrayFieldEditor/
PrimitiveInput/DocRefCatalogPickerField/File/Image, scope=System). Заодно исправлено:
primitive-типы и enum теперь резолвятся через primitiveTypeDef/enumTypeDef (общий
стек — суперсет; каталог этого пока не делает — отдельное улучшение). Пикер базы
временно локализован в странице (единый пикер — шаг 2 #73). systemFields.tsx удалён.

Part of #73.
